refactor: 기존 feed 파일 제거
