+++ src/pages/feed/FeedList.tsx
@@ -1,21 +1,30 @@
-'use client';
-
-import { type FeedItemType } from '@/apis/schema/feed';
-import FeedItem, { FeedSkeletonItem } from '@/app/feed/FeedItem';
+import { type FeedVisibilityType, useInfiniteFeedList } from '@/apis/feed';
 import Empty from '@/components/Empty/Empty';
 import { ROUTER } from '@/constants/router';
+import useIntersect from '@/hooks/useIntersect';
+import FeedItem, { FeedSkeletonItem } from '@/pages/feed/FeedItem';
 import { css } from '@styled-system/css';
 
-function FeedList({ data }: { data?: Array<FeedItemType> }) {
-  if (!data)
-    return (
-      <ul className={feedListCss}>
-        <FeedSkeletonItem />
-        <FeedSkeletonItem />
-      </ul>
-    );
+interface Props {
+  activeTab: FeedVisibilityType;
+}
 
-  if (data.length === 0) {
+function FeedList({ activeTab }: Props) {
+  const { data, isLoading, hasNextPage, isFetching, fetchNextPage } = useInfiniteFeedList({
+    visibility: activeTab,
+    size: 5,
+  });
+
+  const list = data?.content?.filter((feed) => feed.recordImageUrl); // 이미지 없는 경우가 있음. 나중에 리팩토링 + 서버와 이야기, FeedItem에 ErrorBoundary 적용해도 좋을 듯.
+
+  const targetRef = useIntersect(async (entry, observer) => {
+    observer.unobserve(entry.target);
+    if (hasNextPage && !isFetching) fetchNextPage();
+  });
+
+  if (!data || isLoading) return <FeedListSkeleton />;
+
+  if (list?.length === 0) {
     return (
       <div className={emptyFeedCss}>
         <Empty
@@ -32,14 +41,23 @@ function FeedList({ data }: { data?: Array<FeedItemType> }) {
 
   return (
     <ul className={feedListCss}>
-      {data.map((feed) => (
-        <FeedItem key={feed.recordId} {...feed} />
-      ))}
+      {list?.map((feed) => <FeedItem key={feed.recordId} {...feed} />)}
+      <div className={observerCss} ref={targetRef} />
     </ul>
   );
 }
 
 export default FeedList;
+
+function FeedListSkeleton() {
+  return (
+    <ul className={feedListCss}>
+      <FeedSkeletonItem />
+      <FeedSkeletonItem />
+      <FeedSkeletonItem />
+    </ul>
+  );
+}
 
 const emptyFeedCss = css({
   display: 'flex',
@@ -53,4 +71,8 @@ const feedListCss = css({
   display: 'flex',
   flexDirection: 'column',
   gap: '32px',
+});
+
+const observerCss = css({
+  height: '1px',
 });
