+++ src/pages/feed/index.page.tsx
@@ -1,6 +1,6 @@
-import FeedSection from '@/app/feed/FeedSection';
 import AppBarBottom from '@/components/AppBarBottom/AppBarBottom';
 import BottomDim from '@/components/BottomDim/BottomDim';
+import FeedSection from '@/pages/feed/FeedSection';
 import AppBar from '@/pages/home/AppBar';
 
 export default function FeedPage() {
